+++ billing.go
@@ -55,6 +55,7 @@ func (c *Client) RetrievePaymentMethodsWithContext(ctx context.Context) (out *Pa
 	return
 }
 
+// getPaymentMethodByPriority returns the PaymentMethodObject associated with the given PaymentMethodPriority.
 func (c *Client) getPaymentMethodByPriority(priority PaymentMethodPriority) (out *PaymentMethodObject, err error) {
 	paymentMethods, err := c.RetrievePaymentMethods()
 	if err != nil {
